Support handover in interleaved reassembly streams
This was a TODO that is now resolved to be on par with the C++
implementation.

+++ src/api/handover.rs
@@ -48,7 +48,7 @@ pub struct HandoverTransmission {
 
 pub struct HandoverOrderedStream {
     pub id: u16,
-    pub next_ssn: u16,
+    pub next_ssn: u32,
 }
 
 pub struct HandoverUnorderedStream {
